refactor: internal 패키지 생성 및 내부 컴포넌트 이관
