--- Near/app/src/main/java/com/alarmy/near/presentation/ui/component/textfield/NearTextLimitedTextField.kt
@@ -1,2 +1,0 @@
-package com.alarmy.near.presentation.ui.component.textfield
-
